Update remove licence validation to check licence refs (#1702)
* Update remove licence validation to check licence refs

https://eaflood.atlassian.net/browse/WATER-4777

As part of the work to implement notifications in the system repo we have a requirement allow the user to remove licences from teh recipients list.

This change updates the validation for the remove licences page. The user can input anything into the text field. We want to check that each licence is valid before displaying the recipients list or creating the download csv.

The licences to exclude are provided and checked against the due returns logs.