Fix `null` total volume when no meter readings submitted for a month
`formatQuantity()` is documented as returning `null` if the quantity provided is `null` or `undefined`. However the guard clause also returns `null` if the quantity is `0`. This causes a problem when looking at months where no meter readings were submitted, as the total volume is displayed as the string 'null'. We therefore amend `formatQuantity()` to specifically guard against `null` or `undefined`, allowing `0` to be returned as the expected string '0'