Update notifications setup initiate session
https://eaflood.atlassian.net/browse/WATER-4777

As part of the work to implement notifications in the system repo we have multiple journeys for notifications.

This change updates ad-hoc returns initiate session to use a shared initiate session.

This change also adds a 'journey' key to session for the notifications setup journey to enable alternative logic for multiple notification types.