Update notifications ad-hoc error handling (#1710)
* Update notifications ad-hoc error handling

As part of the work to migrate the notification logic into the system repo we have a requirement to handle a ad-hoc (single licence) journey.

This change improves the current error handling (if the licence does not exist it should be rendered in the page when submitted).

We have a pattern for putting validation in a corresponding validation service. This  change creates and uses a validator for ad-hoc licences.

This change also removes redundant use of passing the session to the presenter (this was part of the issue causing the submit service to not display the licence ref if it did not exist).