Edit return log single volume routing (#1716)
https://eaflood.atlassian.net/browse/WATER-4852

During discussions about the return log setup journey, we identified an issue with the routing to the single volume page. Currently, users can only access this page if they select 'No' on the meter details provided page. Additionally, the page is incorrectly displayed for both the volume and meter readings routes, even though it should only appear when users are entering volumes.

To fix this, the routing will be updated so that the single volume page always appears when users select volumes instead of meter readings, regardless of their choice on the meter details page.