Support reissue return logs 4 'no returns required' (#1706)
https://eaflood.atlassian.net/browse/WATER-4552

When a user creates a no return required return version the return logs need to be updated. This PR adds in the logic to correctly void the return logs from the start date of the no returns required version up to the current date or up to the start of any future return version.