Add reference code to notification views (#1724)
https://eaflood.atlassian.net/browse/WATER-4900

This change refactors all the view pages for the notifications setup journey.

We have a common view pattern of displaying relevant information above the page title for users.

This change adds the reference code above the page title for all pages on the notifications setup journey.

The reference code is unique to each journey. It is created when the session is initialised and is used to name downloadable file.