Add reminders journey to notifications setup (#1725)
https://eaflood.atlassian.net/browse/WATER-4900

This change updates the notifications presenter to handle the notification type 'reminders'.

Previous work has been done to dynamically handle the journey data. So this change is minimal.

The initiate session is already tested for reminders. The services check if the journey is 'ad-hoc' and handle that journey differently from 'invitations' and 'reminders'.