Update confirm received service
In the service it self, we just needed to update it to reflect the name change of the fetch service.

The key change was in the tests. There was clearly some confusion with references to sessions, when this page actually has nothing to do with the setup session. Plus tests for things the page doesn't actually do (they happen elsewhere).

But the ky change is to take advantage of why we have services and separate fetch modules. It means we can stub the fetch and avoid interacting with the DB.