Fixup the submit submission tests
We update the 'Arrange' part of the tests to better reflect a 'real' scenario. We also only generate a return log to update for those tests that require it.

We merge those tests that require checking the return log was properly updated to reduce the number of DB interactions we depend on.