Fix return log under query notification
https://eaflood.atlassian.net/browse/WATER-4893

We recently added the ability to mark and unmark returns as 'Under query' to our new view return logs page.

What we didn't spot when it was added, is that it follows our 'flash notification' pattern, which doesn't apply in this scenario.

For example, we are viewing a return log that is not under query.

- the user clicks the `Mark as under query` button
- we set the flag on the return log and redirect back to the page
- we display a [notification banner](https://design-system.service.gov.uk/components/notification-banner/) that states _"This return has been marked under query."_
- the user refreshes, or goes away and comes back to the page, the notification is no longer shown

This is our 'flash notification' pattern. However, it does not apply in this scenario. This is because once the page has been refreshed, the _only_ indication there is a query with the return log is the state of the button, which now has the label `Resolve query`.

The intent was that the banner remains, for example, in the same way the banner is always displayed for licences flagged for supplementary billing.

This change updates the functionality to match this expected behaviour.