Fix missing page title in check recipients page
https://eaflood.atlassian.net/browse/WATER-4775

A recent refactoring to this page meant we were no longer passing in a property `pageTitle`. The layout expects this in order to set the title for the page in the browser.

Needing to move the title out of `text:` to its own property meant there was really no need to group the text properties together, hence this change splits them out.

We drop `send:`, which we were thinking might be worth keeping in case we are expected to do something else with the button. But in light of this fix we decided to drop it now, and sort later if the request comes in.