Amend recipients download CSV headers (licence) (#1728)
https://eaflood.atlassian.net/browse/WATER-4776

We were asked to split the recipients by licence number in the download. This means each licence is on a single line. But we hadn't spotted that the column header was still pluralised as `Licences`.

We've also finally realised that though the ask was 'a licence on each line', it's more like 'a return log on each line'.

This is because the information requested in the CSV is at return log level at its lowest. We cannot group return references, start, end, due dates, etc.

So, the same applies to the return log-related headers.