Add role to download recipients CSV & tweak order
https://eaflood.atlassian.net/browse/WATER-4776

In the new recipients CSV file we've created that can be downloaded from our new 'Check the recipients' page, we spotted it is not possible to identify the _role_ of the recipient. For example, the CSV will tell us it is a letter, but not whether it is to the licence holder or returns contact.

So, the key change is to add this information to the CSV file.

With this added, we also believe the ordering and naming can be improved. After adding 'role', currently it would be the following.

```csv
Message type, Message reference, Role
letter, invitations, Licence holder
```

We believe it should be the following.

```csv
Notification type, Message type, Role
invitations, letter, Licence holder
```