Add role to download recipients CSV & tweak order  (#1729)
https://eaflood.atlassian.net/browse/WATER-4776

In the new recipients CSV file we've created that can be downloaded from our new 'Check the recipients' page, we spotted it is impossible to identify the recipient's _contact type_. For example, the CSV will tell us it is a letter but not whether it is to the 'licence holder' or the 'returns to' contact.

So, the key change is to add this information to the CSV file.

With this added, the ordering and naming can also be improved. After adding 'contact type, ' it would currently be the following:

```csv
Message type, Message reference, Contact type
letter, invitations, Licence holder
```

It should be the following.

```csv
Notification type, Message type, Contact type
invitations, letter, Licence holder
```