Add notifications setup cancel page (#1734)
* Add notifications setup cancel page

https://eaflood.atlassian.net/browse/WATER-49001

This change add the cancel page for the notifications setup journey.

The change displays a new view to the user giving them the option to cancel the notifications they are about the send.

This page works for both 'ad-hoc' and other journeys. In the case of an 'ad-hoc' journey we show the licence number the user previously entered. For the other journeys we will show the returns period selected.

This change also handle the submit action and will clear the session and redirect the user to the /manage page.

---------

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>