Fix notifications duplicate licence refs (#1743)
https://eaflood.atlassian.net/browse/WATER-4775

When a "recipient" is fetched the return logs might have multiple occurrences of with the same "licence_ref". When this was is the case we received duplicate "licence_refs" as a string like this: "01/114,01/114".

This change fixes this issue but selecting a distinct return logs by the licence ref.

We do no need to change the download service as we expect multiple "licence_refs" and return log references for the CSV.