Add GOV.UK notify services (#1737)
* Add GOV.UK notify services

https://eaflood.atlassian.net/browse/WATER-4778

As part of the ongoing work to migrate the legacy applications logic into the system repo we have the requirement to handle notifications using GOV.UK Notify in the system application.

This change adds the initial notify services we will commonly need to use [email, letter].

We want to control the response object from notify and "lightly" take what we need. This is for two reasons:
- If the Notify api changes we have our own internal representation
- We (currently) only needs the status and errors from the response

We want to capture the errors from notify and pass them back to the caller to handle the errors (We do not want the server to stop if a notification fails).

The tests give us a good understanding of the responses we can expect from notify to allow the calling functions to handle appropriately.