Use existing title helpers in review pages part 2
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff bill runs

When working on adding support for `two_part_supplementary` bill runs we spotted that the review pages were not using an existing helper to set the bill run title.

Well, we've spotted some more helpers it is not taking advantage of!

This change updates the review bill run to use more existing helpers.