Update FetchChargeVersions to support supplementary
When generating a 2PT supplementary bill run, the user will have selected a financial year based on what `licence_supplementary_year` records existing.

This means when we fetch the charge versions to start matching and allocating, we should only be looking at those connected to licences that have an entry in `licence_supplementary_year` table for the same year.