Point Edit Return button to Check Details page
https://eaflood.atlassian.net/browse/WATER-4908

Currently the “Edit Return” button on the new “View Returns” page points to the legacy code. Ahead of building out the edit return functionality, we want it to point to the new “Check details” page. This involves taking the submission data and loading it into the session so that it’s correctly displayed in “Check details”.