Housekeeping - stub simple DB calls in TPT annual
We're going to add a new check in the TPT annual generate service to match one we'll include when we create the TPT supplementary generate service.

And, as we'll be building the tests for TPT supplementary anew, we'll follow our current practise of stubbing out the simple calls made to the DB to reduce the number of records we're adding to the DB when running the unit tests.

So, we might as well start with some refactoring of the existing tests to do the same!