Copy FetchBillingAccountsService
Straight up going to admit that due to time pressures, this is a bit of tech debt going on the 'credit card'.

We've confirmed this service does exactly what we need it to do for tpt-supplementary. But we don't have the time to refactor it into a shared 'thing'.

Our other excuse is we _know_ there are other things that could be refactored for reuse across the bill runs. Plus, a recent scare with the legacy PRE-SROC supplementary, has put it on the radar for being rebuilt.

Therefore, we believe there is also a case for holding off refactoring until we have a complete picture of billing.

See, told you we have an excuse! ;-)