Add returns notify presenter
https://eaflood.atlassian.net/browse/WATER-4904

As part of the work to integrate with notify we need to transform the previously established recipients list into a list of notifications to send to notify.

This notification presenter transforms a recipient into a notification. It will work our what notify template to use, the email (if it is an email) and the address (if is a letter).

The service to call this presenter will be added in a later change.