Use existing helpers in view bill page (#1756)
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff bill runs

Spotted whilst adding support for two_part_supplementary bill runs. Our view bill page displays the bill run type. When we generated a two-part tariff supplementary bill run, we saw that the type was displayed incorrectly in the view. When we checked we found it was like PR's #1562 and #1747, where a view was not taking advantage of the available helpers.

This change updates the view bill presenter to take advantage of them, which also resolves our bill run type issue!