Add notification setup even presenter
https://eaflood.atlassian.net/browse/WATER-4904

As part of the work to handle notifications in the system code. We need to follow some patterns established by the legacy code.

This change adds the event presenter to structure the data to insert into 'wabs.public.events'.

This presenter will need to use existing legacy concepts such as setting a 'subtype' in order for the legacy reporting pages to continue to work.