Fix notifications setup fetch recipients
https://eaflood.atlassian.net/browse/WATER-4775

The performance of the query has been particularity slow due to a recent change which introduce a distinct check on the return logs.

The previous change was getting a distinct list for all licences, not distinct return logs where the return log was due (amongst our other where conditions)

This change adds a cte to only join relevant the returns logs. This result in the expected response time for our query.