Add Event persist service (#1753)
* Add Event persist service

https://eaflood.atlassian.net/browse/WATER-4904

As part of the work to handle notifications in the system code. We need to follow some patterns established by the legacy code.

We use 'wabs.public.events' for handling notifications. This table is used to link notifications to 'wabs.public.scheduled_notifications'.

The event table is used on the 'notifications/report' page to show all the notifications.

We need to match the existing structure to allow the legacy code to still function as expected.

This change adds the initial service to create a new event suitable for a returns notification.