Update transactionsMatch() helper
We use `transactionsMatch()` in standard supplementary billing when comparing previous transactions to those generated as part of the 'in progress' bill run.

With one minor tweak that won't effect standard supplementary, we can also make it work when comparing two-part tariff transactions. All other things being equal, if the different between a previous and current transaction is the volume, then both need to go into the bill run so we can 'bill' the charge difference between them.