Update how we use DetermineReturnsPeriodService
https://eaflood.atlassian.net/browse/WATER-4778

We have multiple places using the 'DetermineReturnsPeriodService'. This was becoming complex to reuse and test (needed to stub the time in multiple places).

This change updates all the existing instances of DetermineReturnsPeriodService with a session object containing all the relevant information (session.determinedReturnsPeriod).

This change has simplified working with a chosen returns cycle.