Move reusable supplementary billing services (#1760)
https://eaflood.atlassian.net/browse/WATER-4201

As we write this description, the long outstanding annual SROC two-part tariff bill runs are being 'sent' in production.

The final part of two-part tariff is to add support for supplementary billing; when changes are made to a licence or the returns previously submitted and billed, we need to generate supplementary two-part tariff bill runs.

We've already made several changes to support it, off the back of a [spike](https://github.com/DEFRA/water-abstraction-system/pull/1412).

That spike has shown us there are existing billing services, especially in supplementary billing, that we can reuse for two-part supplementary. This change moves them to make their reuse more transparent, while tweaking a few things to allow reuse.