Add generate TPT supplementary endpoint from review
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff supplementary bill runs

Once a user has finished reviewing a two-part tariff supplementary bill run, they will want to progress it to the next stage, which is where the bills and transactions are generated.

Currently, the `Continue` button in the review bill run screen is hard coded to hit the endpoint to generate an annual two-part tariff bill run.

For now, we want to keep the bill runs separate, so in this change we add a new endpoint that when hit, will trigger generating the TPT supplementary bill run (for now though it'll be a shell service).

This also means making the `Continue` button dynamic by updating the review bill run presenter so that it generates the link.