Add stub service for tpt supplementary
This is where all the work of a billing engine happens. That change is out of scope for what we are doing in this change. So, we'll fill it out later.

For now, this allows us to confirm that our changes to the generate service are working as expected, and it knows which process billing period service to call.