SHow credits/debits in view bill for tpt supp.
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff supplementary bill runs

Spotted whilst adding support for `two_part_supplementary` bill runs. Our view bill page displays credit and debit totals only if the bill run type is 'supplementary'. With us adding two-part tariff supplementary bill runs, we also need it to display them for this bill type.

Because there are a couple of presenters that need this, and we've managed to reduce inconsistencies by ensuring our presenters are using the billing helpers we've created, this change moves the logic to a helper, then updates the views to use it.

Plus, we ensure the helper says 'yes' to displaying credits for two-part tariff supplementary! 😁