Show credits/debits in view bill for tpt supp. (#1766)
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff supplementary bill runs

Spotted whilst adding support for `two_part_supplementary` bill runs. Our view bill page only displays credit and debit totals if the bill run type is 'supplementary'. With us adding two-part tariff supplementary bill runs, we also need it to display them for this bill type.

Because a couple of presenters need this, and we've managed to reduce inconsistencies by ensuring our presenters are using the billing helpers we've created, this change moves the logic to a helper.

But rather than add it to our `BasePresenter`, which is getting a little overloaded, we instead create a new `BillingPresenter`, and move our existing 'billing' helpers to it.

_Then_ we add the helper to this new presenter!

Finaly, we ensure the helper says 'yes' to displaying credits for two-part tariff supplementary! 😁