Housekeeping - stub the insert queries
We just wanted to check that the things we expected to be persisted were, and that they were linked together. When the tests were originally written we were happy to interact with the DB.

With our new ethos, if and when we can we stub the interaction instead. As we're about to duplicate a lot of the testing for supplementary, now seemed a good time to stop interacting with the DB in the existing tests.