Splitting single volume on return log lines (#1733)
https://eaflood.atlassian.net/browse/WATER-4853

As part of the work for the final page on the return log setup journey, a user can enter a single volume and an abstraction period for which the single volume was taken in.

With the pages built all that's left is for the backend to split this single volume amongst the lines that the abstraction period covers on the return.

When a total volume is divided across multiple lines, some of the amounts may have long decimal places. If you add them all back together, the total can end up slightly higher than the original due to small rounding differences. To make sure the total stays correct, rounding is only applied to the last line.