Feature notifications setup submit check (#1755)
* Add Notification setup submit check service
https://eaflood.atlassian.net/browse/WATER-4904

As part of the work to migrate the returns notifications from the legacy code. We need to have a service to send notification to notify.

This change adds the initial service to orchestrate creating, sending and persisting notifications.

The create event and event presenter have been implemented in earlier work and plumbed in this change.


---------

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>