Performance tweak to the existing unflag unbilled
When developing the unit tests we debugged the query Objection was generating, and spotted that it was using `SELECT *` where we were using `relatedQuery()`.

Ordinarily that would be fine, but as we are using it in an `exists` clause it is unnecessary. This change also means it is easier to see the queries are checking for the same things.