Update notifications setup
https://eaflood.atlassian.net/browse/WATER-4778

Work has been done previously to implement the service and presenters. This was done to keep changes and code reviews concise.

A result of this has been some wrong assumptions i.e an array of licences stored as a jsonb filed needed to be converted to a json string prior to saving in the database.

This change starts stitching the service together to create an event and start the notification process (just the presenter, the other services will be developed later).

This change also rectifies some initial assumptions and makes sure our saved data matches the legacy schema.