Handle TPT licence flags during supp process - pt1 (#1771)
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff supplementary bill runs

In recent changes we've been moving and updating existing billing services to allow us to create [two-part tariff supplementary bill runs](https://github.com/DEFRA/water-abstraction-system/pull/1748).

But like standard supplementary, there is another facet the billing engine has to handle; flagging.

When a licence is changed, for example, its charge or return versions, or a return submission, the system will 'flag' the licence for supplementary billing.

This is a boolean field on the licence record for 'standard' supplementary. For a two-part tariff, we add an entry to `water.licence_supplementary_years`.

> N.B. We aim to do the same for 'standard' supplementary in the future.

But the result is the same.

- Being flagged determines which licences are picked up for processing when a supplementary bill run is created
- When a bill run is 'processed', the engine needs to unflag licences where a bill was not found to be needed
- If a licence is removed from the supplementary bill, the flag needs to remain
- Once the bill run is sent, all billed licences should be unflagged
- Those licences that are changed after the bill run is created, or moved into 'workflow' should not be unflagged

The benefit of the `licence_supplementary_year` entries is that we can assign them to a bill run when it is set up. This means we can know more accurately which licences should be unflagged when a bill run is being 'processed'.

This is part one of the changes and covers the following actions from the UI.

- Assigning to a bill run when the bill run is set up
- Removing a licence from 'review'
- Unflagging unbilled licences when a bill run is continued from 'review'
- Removing a bill from a 'ready' bill run
- Removing a licence from a 'ready' bill run

The scenarios around changing licences after the bill run has been set up or added to workflow are also covered. The one step left to complete is when the bill run is 'sent', at which point all assigned licence supplementary year entries need to be removed. That'll come in part 2, but merging part 1 allows our QA team to get started testing this complex functionality.