Handle TPT licence flags during supp process - pt2
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff supplementary bill runs

In our previous change [Handle TPT licence flags during supp process - pt1](https://github.com/DEFRA/water-abstraction-system/pull/1771) we added all the logic for dealing with two-part tariff supplementary billing flags (the `water.licence_supplementary_years` table) whilst the bill run is being setup, created and processed.

The one part we didn't add, was when the bill run is finally 'sent' and we need to unflag all licences that were billed as part of the bill run.

This is that final change.