Do not include 2PT supp. licences already assigned
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff supplementary bill runs

When writing up the ticket to cover the changes made, we realised we'd made a mistake with how we were fetching the charge versions when matching and allocating for a two-part tariff supplementary licence.

We'd updated `app/services/bill-runs/match/fetch-charge-versions.service.js` to handle the 2 two-part tariff bill run types. But when we added the where clause for supplementary, we had forgotten to ensure it ignores licences already assigned to a bill run.

In reality, you should be able to setup a two-part tariff supplementary and assign a licence, and then create another bill run for that region whilst the first is in progress.

But we figured it was better to be safe than sorry! Plus, it makes the intention clearer.