Add notifications setup create notification service (#1772)
* Add notifications setup create notification service

https://eaflood.atlassian.net/browse/WATER-4778

As part of the ongoing work to replace the legacy notifications code. We need to persist notification into 'water.scheduled_notification'.

This requires an adherence to the existing schema.

This change introduces the create returns notification service to persist notifications into the database.

* Add notifications setup create notification service

https://eaflood.atlassian.net/browse/WATER-4778

As part of the ongoing work to replace the legacy notifications code. We need to persist notification into 'water.scheduled_notification'.

This requires an adherence to the existing schema.

This change introduces the create returns notification service to persist notifications into the database.

---------

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>