Add notifications setup scheduled notification presenter
https://eaflood.atlassian.net/browse/WATER-4778

As part of the work to move the returns notifications away from the legacy code we need follow some existing conventions and structure.

This change transforms a 'notification' into a 'scheduledNotifications'. This is required to store a 'scheduledNotification' that matches the existing schema and still work with the legacy code.

Some filed present in 'water.scheduled_notifications' are not used for returns notification they are not populated.

'water.scheduled_notifications.licences' is populated in the legacy code, but it is not used (as far as we can see) so will be omitted.