Fetch only relevant previous transactions in supp.
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff supplementary bill runs

As part of [Move reusable supplementary billing services](https://github.com/DEFRA/water-abstraction-system/pull/1760) we moved `FetchPreviousTransactionsService` so it could be reused by both the standard and two-part tariff supplementary billing engines.

We just assumed on face value that we would fetch the previous transactions for a billing period, let the transactions compare and cancel out as normal, and any credits and debits that shake out at the end are what make up the bill run.

But testing straight away threw up an issue with this assumption.

Imagine the billing period is 2022-23 and we have licence `01/123`. A transaction was added to the annual bill run for 214 billable days and the authorised volume of 18.14ML. It also adds a compensation charge transaction. This is the 'first part charge'.

Then the two-part tariff annual bill run is created for the same period. To that a transaction is added for 0 billable days and the authorised volume of 18.14ML. This is the 'second part charge'.

A change is then made in the period, so the licence is flagged for two-part tariff supplementary. After the 2PT supplementary bill run is generated a user in the review pages allocates 30ML to the entry, then continues the bill run.

At this point the fetch previous before this change was finding 3 debit transactions.

| Bill run | Days | Volume | Charge |
|----------|------|--------|--------|
| Annual   |  214 | 18.184 | 256.50 |
| Annual   |  214 | 18.184 |   0.00 |
| 2PT Anl. |    0 | 18.184 | 256.50 |

None of them match off against each other. So, they were then compared against the transaction generated for the two-part tariff supplementary.

| Bill run | Days | Volume | Charge |
|----------|------|--------|--------|
| 2PT supp |    0 |     30 | 256.50 |

None of the previous is a match, so no lines get cancelled. The end result is a credit of £513 against a debit of £256, with all 4 lines on it.

The problem is `FetchPreviousTransactionsService` now needs to be 'billing specific'. When fetching previous transactions for a standard supplementary bill run, it should only be looking at annual and supplementary bill runs. For a two-part tariff bill run, it should only be looking at previous 2PT annual and supplementary.

With this change, in our scenario the fetch will return only this transaction.

| Bill run | Days | Volume |
|----------|------|--------|
| 2PT Anl. |    0 | 18.184 |

It still wouldn't match to the one generated, but that is expected because the result needs to be the following.

| Bill run | Days | Volume |
|----------|------|--------|
| 2PT Anl. |    0 | 18.184 |
| 2PT supp |    0 |     30 |