Add new previous billing data seeder
After reviewing the current unit tests, we realised they were very much focused on specifics of the query, i.e. if the licence ID passed in has no bill licences, assert we return no results.

To ensure we are fetching the _right_ previous transactions for the supplementary bill run being created, its purpose would be better expressed and tested via scenarios.

We also wanted to limit the amount of records we were creating in the DB. So, we've created this seeder that generates the environment we are trying to deal with.

We have 1 billing account with 2 licences. Both licences have been included in a standard and two-part tariff annual bill run. One of the licences has also been included in a standard and two-part tariff supplementary bill run.

When testing the `FetchPreviousTransactionsService`, if the scenario is we are creating a two-part tariff supplementary bill run, then it should only return the 'two-part-tariff' transactions for each licence when called. The standard transactions must be ignored, and vis-versa.

We'll use the seeder to create _all_ the records, and then fire all our unit tests to ensure it is only returning the ones we expect.

For the first time, we'll also clean up the data generated by the seeder. We name the method `zap()`, so tha we can place it alphabetically below `seed()` in the module!