Move FetchPreviousTransactions out of process
We need to be able to tell fetch the type of bill run we are generating. We could pass it through as another flag via `ProcessSupplementaryTransactionsService`, but it would be another thing we're passing through, just for it to pass through to FetchPrevious.

Better, if we call FetchPrevious from the service that _knows_ the bill run type being generated, and just pass the result through to `ProcessSupplementaryTransactionsService`.