Add notifications setup batch notifications service (#1775)
https://eaflood.atlassian.net/browse/WATER-4778

This change implements the initial service to batch send / save returns notifications.

The batch mechanising will be added in a later change.

This change focuses on putting the initial structure in place to convert recipients to notifications and send them to notify.

The legacy code uses a 'sent' and 'error' count when rendering in the UI. We have matched that existing pattern.